Use native functions for getting terminal data
Termion does not work on Windows, so drop Termion and use the Linux API
and Windows API natively instead. This is done with the libc crate on
Linux and the windows crate on Windows.

This fixes errors upon compiling for Windows and also allows color when
using Invader on Command Prompt which explicitly requires colors to be
turned on via the Windows API where Windows Terminal has them on
already.