Add Cargo.lock to .gitignore
Invader is both a library and executable. Rust recommends to not check
in Cargo.lock if a library and to do so if an executable. This is a bit
conflicting, and Cargo.lock clutters the version history a bit.

If breaking changes do occur with a dependency update, it can probably
just be fixed.