Add more tag reference path checks
Dedupe path separators. Extraneous path separators don't really cause
issues, but they, like uppercase characters, should be silently fixed.

Disallow using "." and ".." as a directory. These cannot be sanely
handled, so it is OK to error on these. This will come at the cost of
Invader not being able to open tags with paths consisting of these, but
a few minutes of manually fixing these tags is worth totally preventing
the security risk of allowing any tag downloaded off the Internet to
point to or extract to tags outside of the tags directory. Since these
are rare to nonexistent, it doesn't matter anyway.

Disallow using paths consisting of only path separators. The HEK can
sometimes resolve these to empty paths but not always, thus they aren't
safe to use and should be considered invalid.