Refactor; rename library to "ringhopper"
CE verb code should be moved to the command line tool and kept out of
the library.

Also, rename the library to "ringhopper" so the Invader command-line
tool can be properly named "invader". Ringhopper will contain the bulk
of the code, but all of the verb code will be in Invader.